+++ src/main/java/com/wanted/naeil/domain/community/controller/PostController.java
@@ -5,7 +5,7 @@ import com.wanted.naeil.domain.community.dto.request.PostCreateRequest;
 import com.wanted.naeil.domain.community.dto.request.PostUpdateRequest;
 import com.wanted.naeil.domain.community.dto.response.PostDetailResponse;
 import com.wanted.naeil.domain.community.dto.response.PostListResponse;
-import com.wanted.naeil.domain.community.entity.PostCategory;
+import com.wanted.naeil.domain.community.entity.enums.PostCategory;
 import com.wanted.naeil.domain.community.service.PostService;
 import com.wanted.naeil.domain.course.entity.Course;
 import com.wanted.naeil.domain.learning.entity.Enrollment;
