+++ src/main/java/com/wanted/naeil/domain/payment/controller/CartController.java
@@ -1,10 +1,11 @@
 package com.wanted.naeil.domain.payment.controller;
 
+import com.wanted.naeil.domain.auth.model.dto.AuthDetails;
 import com.wanted.naeil.domain.payment.dto.response.CartPageResponse;
 import com.wanted.naeil.domain.payment.service.CartService;
 import com.wanted.naeil.domain.payment.service.PaymentService;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -21,31 +22,37 @@ public class CartController {
 
 
     // 로그인 사용자 확인
-    private Long getLoginUserId(HttpSession session) {
-        Long loginUserId = (Long) session.getAttribute("loginUserId");
-        if (loginUserId == null) {
+    private Long getLoginUserId(AuthDetails authDetails) {
+        if (authDetails == null) {
             throw new IllegalStateException("로그인이 필요합니다.");
         }
-        return loginUserId;
+        return authDetails.getLoginUserDTO().getUserId();
     }
 
     // 장바구니 담기
     @PostMapping("/add/{courseId}")
-    public String addCartItem(@PathVariable Long courseId, HttpSession session) {
-        Long loginUserId = getLoginUserId(session);
+    public String addCartItem(@PathVariable Long courseId,
+                              @AuthenticationPrincipal AuthDetails authDetails) {
 
-        cartService.addCartItem(loginUserId, courseId);
+        Long loginUserId = getLoginUserId(authDetails);
 
-        // 강의 상세 페이지로 다시 이동
-        // 프로젝트 실제 상세 경로에 맞게 바꾸면 됨 -> 추후에 확인 필요!!!
-        return "redirect:/courses/" + courseId;
+        try {
+            cartService.addCartItem(loginUserId, courseId);
+            return "redirect:/course/" + courseId + "?cartStatus=added";
+        } catch (IllegalStateException e) {
+            // 강의 상세 페이지로 다시 이동
+            // TODO: 프로젝트 실제 상세 경로에 맞게 바꾸면 됨 -> 추후에 확인 필요!!!
+            return "redirect:/course/" + courseId + "?cartStatus=duplicate";
+        }
     }
 
     // 장바구니 조회(장바구니 메인)
     @GetMapping
-    public String getCartPage(@RequestParam(required = false) Long selectedCartItemId, HttpSession session, Model model) {
+    public String getCartPage(@RequestParam(required = false) Long selectedCartItemId,
+                              @AuthenticationPrincipal AuthDetails authDetails,
+                              Model model) {
 
-        Long loginUserId = getLoginUserId(session);
+        Long loginUserId = getLoginUserId(authDetails);
 
         CartPageResponse cartPage = cartService.getCartPage(loginUserId, selectedCartItemId);
         model.addAttribute("cartPage", cartPage);
@@ -55,23 +62,26 @@ public class CartController {
 
     // 장바구니 삭제 -> @deleteMapping 으로 변경 예정
     @DeleteMapping("/delete/{cartItemId}")
-    public String removeCartItem(@PathVariable Long cartItemId, HttpSession session) {
-        Long loginUserId = getLoginUserId(session);
+    public String removeCartItem(@PathVariable Long cartItemId,
+                                 @AuthenticationPrincipal AuthDetails authDetails) {
+
+        Long loginUserId = getLoginUserId(authDetails);
 
         cartService.removeCartItem(loginUserId, cartItemId);
 
         return "redirect:/cart";
     }
 
+
     // 체크된 장바구니 항목 결제
     @PostMapping("/checkout")
-    public String checkoutCartItems(@RequestParam List<Long> selectedCartItemIds, HttpSession session) {
+    public String checkoutCartItems(@RequestParam List<Long> selectedCartItemIds,
+                                    @AuthenticationPrincipal AuthDetails authDetails) {
 
-        Long loginUserId = getLoginUserId(session);
+        Long loginUserId = getLoginUserId(authDetails);
 
         paymentService.checkoutSelectedCartItems(loginUserId, selectedCartItemIds);
 
         return "redirect:/my-courses";
     }
-
 }
