+++ src/main/java/com/wanted/naeil/domain/payment/controller/PaymentController.java
@@ -1,14 +1,15 @@
 package com.wanted.naeil.domain.payment.controller;
 
+import com.wanted.naeil.domain.auth.model.dto.AuthDetails;
 import com.wanted.naeil.domain.payment.dto.request.SubscriptionPaymentRequest;
 import com.wanted.naeil.domain.payment.service.CartService;
 import com.wanted.naeil.domain.payment.service.PaymentService;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+        import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
@@ -18,10 +19,19 @@ public class PaymentController {
     private final CartService cartService;
     private final PaymentService paymentService;
 
+    private Long getLoginUserId(AuthDetails authDetails) {
+        if (authDetails == null) {
+            throw new IllegalStateException("로그인이 필요합니다.");
+        }
+        return authDetails.getLoginUserDTO().getUserId();
+    }
+
     // 코스 상세에서 '바로결제' -> 해당 코스를 장바구니에 담고 장바구니 페이지로 이동
     @PostMapping("/courses/direct/{courseId}")
-    public String directCoursePayment(@PathVariable Long courseId, HttpSession session) {
-        Long loginUserId = getLoginUserId(session);
+    public String directCoursePayment(@PathVariable Long courseId,
+                                      @AuthenticationPrincipal AuthDetails authDetails) {
+
+        Long loginUserId = getLoginUserId(authDetails);
 
         Long cartItemId = cartService.addOrGetCartItem(loginUserId, courseId);
 
@@ -31,8 +41,9 @@ public class PaymentController {
     // 장바구니에서 선택한 코스 결제
     @PostMapping("/courses")
     public String checkoutSelectedCourses(@RequestParam List<Long> selectedCartItemIds,
-                                          HttpSession session) {
-        Long loginUserId = getLoginUserId(session);
+                                          @AuthenticationPrincipal AuthDetails authDetails) {
+
+        Long loginUserId = getLoginUserId(authDetails);
 
         paymentService.checkoutSelectedCartItems(loginUserId, selectedCartItemIds);
 
@@ -42,9 +53,9 @@ public class PaymentController {
     // 구독권 구매
     @PostMapping("/subscriptions")
     public String subscribe(@ModelAttribute SubscriptionPaymentRequest req,
-                            HttpSession session) {
+                            @AuthenticationPrincipal AuthDetails authDetails) {
 
-        Long loginUserId = getLoginUserId(session);
+        Long loginUserId = getLoginUserId(authDetails);
 
         paymentService.subscribe(loginUserId, req);
 
@@ -54,21 +65,12 @@ public class PaymentController {
     // 구독 자동결제 on/off 변경
     @PatchMapping("/subscriptions/auto-renew")
     public String updateSubscriptionAutoRenew(@RequestParam Boolean autoRenew,
-                                              HttpSession session) {
+                                              @AuthenticationPrincipal AuthDetails authDetails) {
 
-        Long loginUserId = getLoginUserId(session);
+        Long loginUserId = getLoginUserId(authDetails);
 
         paymentService.updateSubscriptionAutoRenew(loginUserId, autoRenew);
 
         return "redirect:/my-page";
     }
-
-    private Long getLoginUserId(HttpSession session) {
-        Long loginUserId = (Long) session.getAttribute("loginUserId");
-        if (loginUserId == null) {
-            throw new IllegalStateException("로그인이 필요합니다.");
-        }
-        return loginUserId;
-    }
-
 }
